Refactor authorisation code into a module
* The hope is that grouped related code will make these responsibilities easier to manage in proximity to others.
